[Sparc][Leon] Corrected supported atomics size for processors supporting Leon CASA instruction back to 32 bits.
This was erroneously checked-in for 64 bits while trying to find if there was a way to get 64 bit atomicity in Leon processors. There is not and this change should not have been checked-in. There is no unit test for this as the existing unit tests test for behaviour to 32 bits, which was the original intention of the code.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@280710 91177308-0d34-0410-b5e6-96231b3b80d8